Implement SAGA pattern for payment processing and event handling in Payment Service.

+++ nigar-payment-service/Models/PaymentFailedEvent.cs
@@ -1,7 +1,8 @@
 namespace nigar_payment_service.Models;
 
-public class PaymentFailedEvent
-{
-    public int ReservationId { get; set; }
-    public string Reason { get; set; }
-}
+
+    public class PaymentFailedEvent
+    {
+        public long ReservationId { get; set; }
+        public string Reason { get; set; }
+    }
